status 404

+++ controllers/locationController.js
@@ -22,6 +22,10 @@ async function createLocation(req,res){
 
 async function updateLocation(req,res){
     try {
+        const location = await Location.findByPk(req.params.id);
+        if(!location){
+            return res.status(404).json({message: 'Location not found'});
+        }
         await Location.update(req.body, { where: { id: req.params.id } });
         res.status(200).end();
     } catch (error) {
@@ -31,6 +35,10 @@ async function updateLocation(req,res){
 
 async function deleteLocation(req,res){
     try {
+        const location = await Location.findByPk(req.params.id);
+        if(!location){
+            return res.status(404).json({message: 'Location not found'});
+        }
         await Location.destroy({ where: { id: req.params.id } });
         res.status(200).end();
     } catch (error) {
